requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
APScheduler==3.10.4
ast-comments==1.2.2
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
black==23.11.0
cachetools==5.5.2
ccxt==4.4.73
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==7.1.2
colorama==0.4.6
coverage==7.8.0
cryptography==44.0.2
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
Faker==20.1.0
fastapi==0.105.0
filelock==3.18.0
flake8==6.1.0
Flask==2.0.1
freqtrade==2024.7
freqtrade-client==2025.3
frozendict==2.4.6
frozenlist==1.5.0
ghp-import==2.1.0
grafana-api==1.0.3
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.25.2
humanize==4.12.2
hypothesis==6.92.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
isort==5.13.2
itsdangerous==2.0.1
janus==2.0.0
Jinja2==3.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.0.350
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lxml==5.3.2
Mako==1.3.9
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==1.4.1
mkdocs-material==9.5.2
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.0
multidict==6.3.2
multitasking==0.0.11
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.24.4
openai==1.3.7
orjson==3.10.16
packaging==23.2
paginate==0.5.7
pandas==2.2.3
pandas_ta==0.3.14b0
passlib==1.7.4
pathspec==0.12.1
peewee==3.17.9
platformdirs==4.3.7
pluggy==1.5.0
pre-commit==3.5.0
prometheus-client==0.19.0
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
psycopg==3.1.18
psycopg-binary==3.1.18
py_find_1st==1.1.7
pyarrow==19.0.1
pyasn1==0.6.1
pycares==4.6.0
pycodestyle==2.11.1
pycoingecko==3.2.0
pycparser==2.22
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyflakes==3.1.0
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.7.1
pymongo==4.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-rapidjson==1.20
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.2
pyyaml_env_tag==0.1
questionary==2.1.0
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
schedule==1.2.2
scikit-learn==1.3.2
scipy==1.15.2
sdnotify==0.3.2
sentry-sdk==1.39.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.27.0
ta-lib @ file:///C:/Users/panto/Downloads/ta_lib-0.6.3-cp311-cp311-win_amd64.whl#sha256=c3a4e29e7e97dbd03c8d8bef0d26da3ced6cc34cb6af56134d273a0284c748d8
tabulate==0.9.0
technical==1.5.0
tenacity==8.2.3
threadpoolctl==3.6.0
tiktoken==0.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.24.0
virtualenv==20.30.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==2.0.1
xgboost==1.7.6
yarl==1.19.0
yfinance==0.2.33
alpaca-trade-api>=2.3.0
pandas>=1.3.5
numpy>=1.21.0
matplotlib>=3.5.1
python-dotenv>=0.20.0
aiohttp>=3.8.1
asyncio>=3.4.3
